[Fix] int32/int16/int8 in range partition
Support int32/int16/int8 as partition key in range partition
Link: https://code.alibaba-inc.com/oceanbase/obkv-table-client-java/codereview/12763931
* [Fix] int32/int16/int8 in range partition

+++ src/test/java/ci.sql
@@ -113,7 +113,7 @@ PRIMARY KEY (`c1`, `c2`, `c3`, `c4`)
 partition by key(`c1`, `c2`, `c3`) subpartition by key(`c4`) subpartitions 4 partitions 16;
 
 CREATE TABLE IF NOT EXISTS `testPartitionRangeComplex` (
-    `c1` bigint NOT NULL,
+    `c1` int NOT NULL,
     `c2` bigint NOT NULL,
     `c3` varbinary(1024) NOT NULL,
     `c4` varchar(1024) NOT NULL,
